Add enum data to distinguish between ocorrencia and taxa delito

+++ src/main/java/com/umfrancisco/estatisticas_criminais_project/repository/TaxaDelitoRepository.java
@@ -3,6 +3,6 @@ package com.umfrancisco.estatisticas_criminais_project.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.umfrancisco.estatisticas_criminais_project.model.TaxaDelito;
 
-public interface TaxaDelitoRepository extends JpaRepository<TaxaDelito, Integer> {
+public interface TaxaDelitoRepository extends JpaRepository<TaxaDelito, Long> {
 
 }
